Module cryptopals::challenges

Solutions to the cryptopals challenges themselves, broken down by set.

Each challenge returns a ChallengeResults. These are the results of a particular challenge, including, when relevant, encrypted/decrypted inputs/outputs, cryptographic keys, outcomes of attacks and other such data. These outputs will be printed to the terminal when the challenge is run as part of the executable, and are also checked against the known answers as unit tests for the challenges.

Modules

helpers
Helper functions which are useful in solving the challenges but too large to sensibly be inline in the solution itself.

set1
Solutions to the challenges in Set 1.

set2
Solutions to the challenges in Set 2.

Structs

ChallengeResults
The results of running a particular challenge.

ChallengeResultsBuilder
A builder for the ChallengeResults struct.