Struct cryptopals::challenges::ChallengeResults

pub struct ChallengeResults {
    pub set: usize,
    pub challenge: usize,
    pub description: String,
    pub outputs: Vec<(String, String)>,
}

The results of running a particular challenge.

Fields
set: usize

The set that this challenge belongs to.

challenge: usize

The challenge number.

description: String

A description of the challenge.

outputs: Vec<(String, String)>

An array holding the outputs from this challenge.

Methods

impl ChallengeResults

fn check(&self, exp_key: &str, exp_val: &str)

Assert that a particular key is present and holds the expected value.

fn check_prefix(&self, exp_key: &str, exp_val: &str)

Assert that a particular key is present and starts with the given prefix.