Struct cryptopals::utils::data::Data

pub struct Data {
    // some fields omitted
}

Structure which holds the contents of a message.

Methods

impl Data

fn new() -> Data

Creates a new empty Data object.

fn random(size: usize) -> Data

Create a new random Data object containing the given number of bytes.

fn from_hex(input: &str) -> Result<Data, FromHexError>

Creates a new Data object from a sequence of bytes given as a hexadecimal string.

fn from_base64(input: &str) -> Result<Data, FromBase64Error>

Creates a new Data object from a base-64 encoded string.

fn from_text(input: &str) -> Data

Creates a new Data object from a sequence of byte values represented as a plain text string.

fn from_bytes(input: Vec<u8>) -> Data

Creates a new Data object from a sequence of raw byte values.

fn from_single_byte(input: u8) -> Data

Creates a new Data object representing a single byte.

fn to_hex(&self) -> String

Returns the message as a hexadecimal string.

fn to_base64(&self) -> String

Returns the message as a base-64 encoded string.

fn to_text(&self) -> String

Returns the message as a plain text string.

fn bytes(&self) -> &[u8]

Returns a slice containing the sequence of bytes stored in this Data

fn slice(&self, start: usize, end: usize) -> Data

Returns a new Data formed of a slice of the sequence of bytes stored in this Data.

fn len(&self) -> usize

Returns the length (in bytes) of this Data.

fn is_empty(&self) -> bool

Returns true if this Data is empty, and false otherwise.