Module cryptopals::utils::metrics

Functions for evaluating pieces of data according to various metrics.

Enums

HammingDistanceError
Errors that can arise when calculating the Hamming distance between two pieces of data.

Functions

hamming_distance
Calculates the Hamming distance between two pieces of data of equal size.

has_repeated_blocks
Returns true or false as to whether the given data has any repeated blocks.

score_as_english
Returns a numeric score representing how likely it is that a particular piece of data is English text - a higher score is better.

score_xor_keysize
Returns a numeric score representing how likely it is that the given data has been encoded using a repeating XOR key of the given size - lower is better.