Struct cryptopals::victims::block::CbcCookie

pub struct CbcCookie {
    // some fields omitted
}

Encrypts user-provided data in the form of a cookie using CBC.

A black box which takes an arbitrary string as input, escapes any metacharacters (';' and '=') and produces a cookie-like structure of the form comment1=cooking%20MCs;userdata=<user-data>;comment2=%20like%20a%20pound%20of%20bacon. This is then encrypted using CBC mode and provided to the user.

Goal

To obtain, by any means, an encrypted cookie which decrypts to one containing ";admin=true;".

Methods

impl CbcCookie

fn new() -> CbcCookie

Creates a new CbcCookie.

fn make_token(&self, user_data: &str) -> Data

Create an encrypted cookie for the given user data.

fn is_admin(&self, token: &Data) -> bool

Parses an encrypted token, and returns true or false according to whether the token represents a profile containing admin=true.