Struct cryptopals::victims::block::EcbUserProfile

pub struct EcbUserProfile {
    // some fields omitted
}

Creates ECB-encrypted user tokens.

A black box which takes an email address, sanitises the input, creates a token of the form email=<user_email>&uid=10&role=user, then encrypts that under ECB using a block size of 16 bytes, and returns the result.

Goal

To obtain (by any means) a token which decrypts to one containing "role=admin".

Methods

impl EcbUserProfile

fn new() -> EcbUserProfile

Creates a new EcbUserProfile.

fn make_token(&self, email: &str) -> Data

Create a token for the given email address.

fn is_admin(&self, token: &Data) -> bool

Parses an encrypted token, and returns true or false according to whether the token represents a profile containing role=admin.